chore: 오탈자 수정 (WR9-68) #97

+++ src/main/java/io/crops/warmletter/domain/timeline/controller/NotificationController.java
@@ -3,7 +3,6 @@ package io.crops.warmletter.domain.timeline.controller;
 import io.crops.warmletter.domain.timeline.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,7 +16,7 @@ public class NotificationController {
     private final NotificationService notificationService;
 
     @GetMapping(value = "/sub/{memberId}", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
-    public SseEmitter subscribeNotification(@PathVariable int memberId) {
+    public SseEmitter subscribeNotification(@PathVariable long memberId) {
         return notificationService.subscribeNotification(memberId);
     }
 }
